change init_src to 0 if manipulating state
pulling out hair why low/high states of nature giving exact same results. realized that for speed we had init_src = 1 for the iterations, which is fine for the base case but we actually want to see what happens when it reads the value from the control. if this doesn't work i don't know what will - this was the only difference i could find between these iters and previous (where state was done externally)